+++ src/Tools/Editor.css
@@ -120,7 +120,7 @@ svg {
   flex-direction: row;
   align-items: center;
   justify-content: space-evenly;
-  gap: 2px;
+  gap: 0.5rem;
   color: #6e0b0b;
 }
